refactor: 팀 문화 데이터 타입 정의 및 상수 수정
