+++ src/constants/teamCulture.ts
@@ -1,11 +1,11 @@
-export const teamCulture = [
+import { Culture } from '@/types';
+
+export const teamCulture: Culture[] = [
   {
     id: '1',
     title: '어쩌구 저쩌구\n입니다.',
     description:
       '텍스트 입력 텍스트 입력 텍스트 입력 텍스트 입력 텍스트 입력 텍스트 입력 텍스트 입력 텍스트 입력 텍스트 입력 텍스트 입력 텍스트 입력 텍스트 입력',
-    textColor: '#ffffff',
-    imageUrl: 'https://picsum.photos/id/4/200/300',
   },
   {
     id: '2',
@@ -24,7 +24,5 @@ export const teamCulture = [
     title: '어쩌구 저쩌구\n입니다.',
     description:
       '텍스트 입력 텍스트 입력 텍스트 입력 텍스트 입력 텍스트 입력 텍스트 입력 텍스트 입력 텍스트 입력 텍스트 입력 텍스트 입력 텍스트 입력 텍스트 입력',
-    textColor: '#ffffff',
-    imageUrl: 'https://picsum.photos/id/3/200/300',
   },
 ];
